feat(mock-library): function complete need to add support for mock library
